5/29
Cart Page now displays the items that were added to the cart. It can be deleted with the corresponding delete button
All the items in the online order page are buttons that send forms to add the specific item to the database.
-A problem is that when item is added to cart, even if it is a duplicate, it adds the item like it is new.
Need to group the same items together and display the qty of that item